recovery.bin: Update to fix rename bug with FAT16
Fix a bug which could cause the rename of recovery.bin to fail if
RECOVERY.000 (or higher number) existed and was the last root
directory entry in a FAT16 partition. This is unlikely be a problem
for most updates because the rpi-eeprom-update package removes the
renamed recovery.bin files when it runs.